feat: replace postgres with sqlite
- Swap Bun.sql (postgres) for bun:sqlite
- Rewrite schema in migrate.ts for SQLite dialect
- Add db wrapper with RETURNING support, array expansion,
  and boolean normalization (0/1 -> true/false)
- Replace gen_random_uuid() with crypto.randomUUID() in JS
- Replace TEXT[] arrays with JSON text + json_each()
- Replace Postgres casts (::int, ::float, ::uuid) with SQLite
  compatible queries
- Replace now() with datetime('now')
- Replace UNNEST with json_each for label reputation decay
- Update all tests to use explicit UUIDs and SQLite syntax
- Add migrate step to systemd deploy service
- Add fastminds.db to .gitignore
- Remove docker-compose.yml (no longer needed)

+++ src/routes/admin.ts
@@ -35,7 +35,7 @@ admin.get("/conversations", adminMiddleware, async (c) => {
       recipient.username AS recipient_username,
       last_messages.last_body,
       last_messages.last_created_at,
-      (SELECT COUNT(*)::int FROM conversation_messages WHERE conversation_id = conversations.id) AS message_count
+      (SELECT COUNT(*) FROM conversation_messages WHERE conversation_id = conversations.id) AS message_count
     FROM conversations
     JOIN posts ON conversations.post_id = posts.id
     JOIN users AS initiator ON conversations.initiator_id = initiator.id
@@ -45,7 +45,7 @@ admin.get("/conversations", adminMiddleware, async (c) => {
     LIMIT ${limit} OFFSET ${offset}
   `;
 
-  const [countRow] = await db`SELECT COUNT(*)::int AS total FROM conversations`;
+  const [countRow] = await db`SELECT COUNT(*) AS total FROM conversations`;
 
   const conversations = rows.map((row) => ({
     id: row.id,
